1. 添加开放接口页面查询功能，包括单次查询、批量查询 2. 添加开放接口账单管理相关功能，包括应用账单、客户账单 3. 添加开放接口计费功能实现，支持按计费规则配置 4. particle-demo前端打包

+++ component/open-platform/open-platform-client/src/main/java/com/particle/openplatform/client/bill/dto/command/OpenplatformProviderRecordPrdApiDaySummaryCreateCommand.java
@@ -5,6 +5,7 @@ import com.particle.common.client.dto.command.AbstractBaseCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -46,7 +47,7 @@ public class OpenplatformProviderRecordPrdApiDaySummaryCreateCommand extends Abs
 
     @NotNull(message = "平均单价金额 不能为空")
         @Schema(description = "平均单价金额",requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
 
     @NotNull(message = "总消费金额 不能为空")
