+++ component/open-platform/open-platform-client/src/main/java/com/particle/openplatform/client/bill/dto/command/OpenplatformProviderRecordPrdMonthBillCreateCommand.java
@@ -5,6 +5,7 @@ import com.particle.common.client.dto.command.AbstractBaseCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 /**
  * <p>
@@ -46,7 +47,7 @@ public class OpenplatformProviderRecordPrdMonthBillCreateCommand extends Abstrac
 
     @NotNull(message = "平均单价金额 不能为空")
         @Schema(description = "平均单价金额",requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
 
     @NotNull(message = "总消费金额 不能为空")
