+++ component/open-platform/open-platform-client/src/main/java/com/particle/openplatform/client/bill/dto/command/representation/OpenplatformProviderRecordPrdApiDaySummaryQueryListCommand.java
@@ -5,6 +5,8 @@ import com.particle.global.light.share.mybatis.anno.Like;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 /**
  * <p>
@@ -41,7 +43,7 @@ public class OpenplatformProviderRecordPrdApiDaySummaryQueryListCommand extends 
 
 
     @Schema(description = "平均单价金额")
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
 
     @Schema(description = "总消费金额")
