+++ component/open-platform/open-platform-client/src/main/java/com/particle/openplatform/client/bill/dto/command/representation/OpenplatformProviderRecordPrdMonthBillPageQueryCommand.java
@@ -4,6 +4,9 @@ import com.particle.global.light.share.mybatis.anno.Like;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 /**
  * <p>
  * 开放平台供应商月账单 通用分页查询指令对象
@@ -39,7 +42,7 @@ public class OpenplatformProviderRecordPrdMonthBillPageQueryCommand extends Abst
 
 
     @Schema(description = "平均单价金额")
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
 
     @Schema(description = "总消费金额")
