+++ component/open-platform/open-platform-client/src/main/java/com/particle/openplatform/client/bill/dto/data/OpenplatformProviderRecordPrdApiMonthSummaryVO.java
@@ -1,5 +1,6 @@
 package com.particle.openplatform.client.bill.dto.data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import com.particle.common.client.dto.data.AbstractBaseIdVO;
 
@@ -37,7 +38,7 @@ public class OpenplatformProviderRecordPrdApiMonthSummaryVO extends AbstractBase
     private Integer totalFeeCall;
     
     @Schema(description = "平均单价金额")
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
     
     @Schema(description = "总消费金额")
     private Integer totalFeeAmount;
