+++ component/open-platform/open-platform-domain/src/main/java/com/particle/openplatform/domain/bill/OpenplatformOpenapiRecordAppOpenapiDaySummary.java
@@ -4,6 +4,8 @@ import com.particle.common.domain.AggreateRoot;
 import com.particle.global.domain.DomainFactory;
 import com.particle.global.domain.Entity;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 /**
  * <p>
@@ -25,7 +27,7 @@ public class OpenplatformOpenapiRecordAppOpenapiDaySummary extends AggreateRoot 
     private Long openplatformAppId;
 
     /**
-    * 应用id
+    * appId
     */
     private String appId;
 
@@ -57,7 +59,7 @@ public class OpenplatformOpenapiRecordAppOpenapiDaySummary extends AggreateRoot 
     /**
     * 平均单价金额，单位分
     */
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
     /**
     * 总消费金额，单位分
