+++ component/open-platform/open-platform-domain/src/main/java/com/particle/openplatform/domain/bill/OpenplatformOpenapiRecordAppOpenapiMonthSummary.java
@@ -4,6 +4,9 @@ import com.particle.common.domain.AggreateRoot;
 import com.particle.global.domain.DomainFactory;
 import com.particle.global.domain.Entity;
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 /**
  * <p>
  * 开放平台应用开放接口月汇总 领域模型
@@ -24,7 +27,7 @@ public class OpenplatformOpenapiRecordAppOpenapiMonthSummary extends AggreateRoo
     private Long openplatformAppId;
 
     /**
-    * 应用id
+    * appId
     */
     private String appId;
 
@@ -61,7 +64,7 @@ public class OpenplatformOpenapiRecordAppOpenapiMonthSummary extends AggreateRoo
     /**
     * 平均单价金额，单位分
     */
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
     /**
     * 总消费金额，单位分
