+++ component/open-platform/open-platform-domain/src/main/java/com/particle/openplatform/domain/bill/OpenplatformProviderRecordPrdApiMonthSummary.java
@@ -4,6 +4,9 @@ import com.particle.common.domain.AggreateRoot;
 import com.particle.global.domain.DomainFactory;
 import com.particle.global.domain.Entity;
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 /**
  * <p>
  * 开放平台供应商接口月汇总 领域模型
@@ -51,7 +54,7 @@ public class OpenplatformProviderRecordPrdApiMonthSummary extends AggreateRoot {
     /**
     * 平均单价金额，单位分
     */
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
     /**
     * 总消费金额，单位分
