+++ component/open-platform/open-platform-infrastructure/src/main/java/com/particle/openplatform/infrastructure/bill/dos/OpenplatformProviderRecordPrdApiDaySummaryDO.java
@@ -7,6 +7,8 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import lombok.Data;
 import lombok.experimental.Accessors;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 /**
  * <p>
@@ -49,7 +51,7 @@ public class OpenplatformProviderRecordPrdApiDaySummaryDO extends BaseDO {
     /**
     * 平均单价金额，单位分
     */
-    private Integer averageUnitPriceAmount;
+    private BigDecimal averageUnitPriceAmount;
 
     /**
     * 总消费金额，单位分
